requirements.txt
aiohttp==3.7.4.post0
aioredis==1.3.1
aniso8601==7.0.0
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.7
cffi==1.14.5
channels==3.0.4
channels-redis==3.2.0
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
dj-database-url==0.5.0
Django==3.2.3
django-channels-graphql-ws==0.8.0
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
django-sendgrid-v5==1.1.0
djangorestframework==3.12.4
graphene==2.1.8
graphene-django==2.12.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
msgpack==1.0.2
multidict==5.1.0
promise==2.3
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-http-client==3.3.2
django-redis
pytz==2021.1
Rx==1.6.1
sendgrid==6.8.0
service-identity==21.1.0
singledispatch==3.6.1
six==1.16.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.1
text-unidecode==1.3
toml==0.10.2
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.10.0.0
Unidecode==1.2.0
whitenoise==5.3.0
yarl==1.6.3
zope.interface==5.4.0
